+++ app/src/main/java/com/cmput301/w19t06/theundesirablejackals/Message.java
@@ -1,5 +1,7 @@
 package com.cmput301.w19t06.theundesirablejackals;
 
+import com.cmput301.w19t06.theundesirablejackals.User.User;
+
 public class Message extends Communication {
     private String text;
     private Boolean seen;
