+++ app/src/main/res/values/colors.xml
@@ -1,6 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <color name="colorPrimary">#008577</color>
-    <color name="colorPrimaryDark">#00574B</color>
-    <color name="colorAccent">#D81B60</color>
+    <color name="colorPrimaryDark">#00574b</color>
+    <color name="colorAccent">#d81b60</color>
+    <color name="tabTextColor">#10100E</color>
+    <color name="tabIndicatorColor">#056fff</color>
+    <color name="tabBackgroundColor">#a39398</color>
 </resources>
