+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,13 @@
 <resources>
     <string name="app_name">TheUndesirableJackals</string>
+    <string name="user_name">User name:</string>
+    <string name="password">Password:</string>
+    <string name="login">Login</string>
+    <string name="hint_username">Username</string>
+    <string name="hint_password">Password</string>
+    <string name="hint_email">Email</string>
+    <string name="hint_phone">Phone number</string>
+    <string name="tv_signup_login_info">Login Information</string>
+    <string name="tv_signup_contact_info">Contact Information</string>
+    <string name="btn_signup_signup_text">Sign up</string>
 </resources>
