+++ app/src/test/java/com/cmput301/w19t06/theundesirablejackals/AuthenticationTest.java
@@ -1,5 +1,7 @@
 package com.cmput301.w19t06.theundesirablejackals;
 
+import com.cmput301.w19t06.theundesirablejackals.User.Authentication;
+
 import org.junit.Before;
 import org.junit.Test;
