+++ app/src/test/java/com/cmput301/w19t06/theundesirablejackals/BookListTest.java
@@ -1,4 +1,9 @@
 package com.cmput301.w19t06.theundesirablejackals;
+import com.cmput301.w19t06.theundesirablejackals.Book.Book;
+import com.cmput301.w19t06.theundesirablejackals.Book.BookList;
+import com.cmput301.w19t06.theundesirablejackals.Book.BookStatus;
+import com.cmput301.w19t06.theundesirablejackals.User.User;
+
 import org.junit.Before;
 import org.junit.Test;
