+++ app/src/test/java/com/cmput301/w19t06/theundesirablejackals/BookTest.java
@@ -1,9 +1,13 @@
 package com.cmput301.w19t06.theundesirablejackals;
 
+import com.cmput301.w19t06.theundesirablejackals.Book.Book;
+import com.cmput301.w19t06.theundesirablejackals.Book.BookGenres;
+import com.cmput301.w19t06.theundesirablejackals.Book.BookStatus;
+import com.cmput301.w19t06.theundesirablejackals.User.User;
+
 import org.junit.Before;
 import org.junit.Test;
 
-import java.lang.reflect.Array;
 import java.util.ArrayList;
 
 import static org.junit.Assert.*;
