+++ app/src/test/java/com/cmput301/w19t06/theundesirablejackals/ContactInformationTest.java
@@ -1,9 +1,9 @@
 package com.cmput301.w19t06.theundesirablejackals;
 
+import com.cmput301.w19t06.theundesirablejackals.User.ContactInformation;
+
 import org.junit.Before;
 import org.junit.Test;
-
-import java.util.ArrayList;
 
 import static org.junit.Assert.*;
