+++ app/src/test/java/com/cmput301/w19t06/theundesirablejackals/UserTest.java
@@ -1,5 +1,11 @@
 package com.cmput301.w19t06.theundesirablejackals;
 
+import com.cmput301.w19t06.theundesirablejackals.Book.Book;
+import com.cmput301.w19t06.theundesirablejackals.Book.BookGenres;
+import com.cmput301.w19t06.theundesirablejackals.Book.BookRequest;
+import com.cmput301.w19t06.theundesirablejackals.Book.BookRequestList;
+import com.cmput301.w19t06.theundesirablejackals.User.User;
+
 import org.junit.Before;
 import org.junit.Test;
